+++ API/src/takyon_socket.c
@@ -56,7 +56,7 @@ GLOBAL_VISIBILITY bool tknSend(TakyonPath *path, int buffer_index, uint64_t byte
   // Validate fits on recver
   uint64_t total_bytes_to_recv = dest_offset + bytes;
   if (total_bytes_to_recv > buffer->remote_max_recver_bytes) {
-    TAKYON_RECORD_ERROR(path->attrs.error_message, "Out of range for the recver. Exceeding by %lld bytes\n", total_bytes_to_recv - buffer->remote_max_recver_bytes);
+    TAKYON_RECORD_ERROR(path->attrs.error_message, "Out of range for the recver. Exceeding by %ju bytes\n", total_bytes_to_recv - buffer->remote_max_recver_bytes);
     return false;
   }
 
@@ -152,13 +152,13 @@ GLOBAL_VISIBILITY bool tknRecv(TakyonPath *path, int buffer_index, uint64_t *byt
     buffer->got_data = false;
     // Verbosity
     if (path->attrs.verbosity & TAKYON_VERBOSITY_SEND_RECV_MORE) {
-      printf("%-15s (%s:%s) Data was waiting: buf=%d, bytes=%lld, offset=%lld\n",
+      printf("%-15s (%s:%s) Data was waiting: buf=%d, bytes=%ju, offset=%ju\n",
              __FUNCTION__,
              path->attrs.is_endpointA ? "A" : "B",
              path->attrs.interconnect,
              buffer_index,
-             (unsigned long long)buffer->recved_bytes,
-             (unsigned long long)buffer->recved_offset);
+             buffer->recved_bytes,
+             buffer->recved_offset);
     }
     return true;
   }
@@ -193,13 +193,13 @@ GLOBAL_VISIBILITY bool tknRecv(TakyonPath *path, int buffer_index, uint64_t *byt
 
     // Verbosity
     if (path->attrs.verbosity & TAKYON_VERBOSITY_SEND_RECV_MORE) {
-      printf("%-15s (%s:%s) Got header: buf=%d, bytes=%lld, offset=%lld\n",
+      printf("%-15s (%s:%s) Got header: buf=%d, bytes=%ju, offset=%ju\n",
              __FUNCTION__,
              path->attrs.is_endpointA ? "A" : "B",
              path->attrs.interconnect,
              recved_buffer_index,
-             (unsigned long long)recved_bytes,
-             (unsigned long long)recved_offset);
+             recved_bytes,
+             recved_offset);
     }
 
     // Get the data
@@ -208,7 +208,7 @@ GLOBAL_VISIBILITY bool tknRecv(TakyonPath *path, int buffer_index, uint64_t *byt
     uint64_t recver_max_bytes = path->attrs.recver_max_bytes_list[buffer_index];
     if (total_bytes > recver_max_bytes) {
       buffers->connection_failed = true;
-      TAKYON_RECORD_ERROR(path->attrs.error_message, "Receiving out of bounds data. Max is %lld bytes but trying to recv %lld plus offset of %lld\n", (unsigned long long)recver_max_bytes, (unsigned long long)recved_bytes, (unsigned long long)recved_offset);
+      TAKYON_RECORD_ERROR(path->attrs.error_message, "Receiving out of bounds data. Max is %ju bytes but trying to recv %ju plus offset of %ju\n", recver_max_bytes, recved_bytes, recved_offset);
       return false;
     }
     if (recver_buffer->got_data) {
@@ -236,12 +236,12 @@ GLOBAL_VISIBILITY bool tknRecv(TakyonPath *path, int buffer_index, uint64_t *byt
       if (offset_ret != NULL) *offset_ret = recved_offset;
       // Verbosity
       if (path->attrs.verbosity & TAKYON_VERBOSITY_SEND_RECV_MORE) {
-        printf("%-15s (%s:%s) Got data: %lld bytes at offset %lld on buffer %d\n",
+        printf("%-15s (%s:%s) Got data: %ju bytes at offset %ju on buffer %d\n",
                __FUNCTION__,
                path->attrs.is_endpointA ? "A" : "B",
                path->attrs.interconnect,
-               (unsigned long long)recved_bytes,
-               (unsigned long long)recved_offset,
+               recved_bytes,
+               recved_offset,
                buffer_index);
       }
       break;
